Feat/resume refactor

+++ src/main/java/org/nexusscode/backend/application/service/ApplicationService.java
@@ -8,8 +8,6 @@ import java.net.URL;
 import java.time.Instant;
 import java.time.LocalDateTime;
 import java.time.ZoneId;
-import java.util.List;
-import java.util.Map;
 import javax.imageio.ImageIO;
 import lombok.RequiredArgsConstructor;
 import net.sourceforge.tess4j.ITesseract;
@@ -21,6 +19,7 @@ import org.nexusscode.backend.application.domain.Status;
 import org.nexusscode.backend.application.dto.ApplicationRequestDto;
 import org.nexusscode.backend.application.dto.ApplicationResponseDto;
 import org.nexusscode.backend.application.dto.ApplicationSimpleDto;
+import org.nexusscode.backend.application.dto.MemoRequestDto;
 import org.nexusscode.backend.application.repository.JobApplicationRepository;
 import org.nexusscode.backend.global.exception.CustomException;
 import org.nexusscode.backend.global.exception.ErrorCode;
@@ -30,9 +29,6 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
-import org.springframework.http.HttpEntity;
-import org.springframework.http.HttpHeaders;
-import org.springframework.http.MediaType;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
@@ -152,32 +148,38 @@ public class ApplicationService {
         }
     }
 
-    public ApplicationResponseDto getApplication(Long applicationId) {
+    public ApplicationResponseDto getApplication(Long userId, Long applicationId) {
+        User user = userService.findById(userId);
         JobApplication application = applicationRepository.findById(applicationId).orElseThrow(
             ()->new CustomException(ErrorCode.NOT_FOUND)
         );
-        // 추후 로그인 유저의 application 맞는지 확인 필요
+        if(application.getUser()!=user){
+            throw new CustomException(ErrorCode.NOT_UNAUTHORIZED_APPLICATION);
+        }
         return new ApplicationResponseDto(application);
     }
 
     @Transactional
-    public void deleteApplication(Long applicationId) {
+    public void deleteApplication(Long userId, Long applicationId) {
+        User user = userService.findById(userId);
         JobApplication application = applicationRepository.findById(applicationId).orElseThrow(
             ()->new CustomException(ErrorCode.NOT_FOUND)
         );
-        // 추후 로그인 유저의 application 맞는지 확인 필요
+        if(application.getUser()!=user){
+            throw new CustomException(ErrorCode.NOT_UNAUTHORIZED_APPLICATION);
+        }
         applicationRepository.delete(application);
     }
 
-    public Page<ApplicationSimpleDto> getAllApplication(int page, int size) {
+    public Page<ApplicationSimpleDto> getAllApplication(Long userId, int page, int size) {
+        User user = userService.findById(userId);
         Pageable pageable = PageRequest.of(page, size);
-        Page<JobApplication> applicationPage = applicationRepository.findAll(pageable);
-        // 추후 로그인 유저의 application 로만 리스트 조회
+        Page<JobApplication> applicationPage = applicationRepository.findAllByUser(user,pageable);
 
         return applicationPage.map(ApplicationSimpleDto::new);
     }
 
-    @Transactional
+    /*@Transactional
     public String uploadDetailImage(Long applicationId, MultipartFile file) {
         JobApplication application = findById(applicationId);
         String imageText;
@@ -199,11 +201,12 @@ public class ApplicationService {
         applicationRepository.save(application);
 
         return imageText;
-    }
+    }*/
 
-    public Page<ApplicationSimpleDto> searchApplication(int page, int size,String searchWord) {
+    public Page<ApplicationSimpleDto> searchApplication(Long userId, int page, int size,String searchWord) {
+        User user = userService.findById(userId);
         Pageable pageable = PageRequest.of(page, size);
-        Page<JobApplication> jobApplicationPage = applicationRepository.findByCompanyNameContainingIgnoreCaseOrJobTitleContainingIgnoreCase(searchWord,searchWord,pageable);
+        Page<JobApplication> jobApplicationPage = applicationRepository.searchByUserAndCompanyOrTitle(user,searchWord,pageable);
 
         return jobApplicationPage.map(ApplicationSimpleDto::new);
     }
@@ -212,5 +215,17 @@ public class ApplicationService {
         return applicationRepository.findById(id).orElseThrow(
             ()->new CustomException(ErrorCode.NOT_FOUND_APPLICATION)
         );
+    }
+
+    @Transactional
+    public void updateMemo(Long userId, Long applicationId, MemoRequestDto memoRequestDto) {
+        User user = userService.findById(userId);
+        JobApplication application = findById(applicationId);
+        if(application.getUser()!=user){
+            throw new CustomException(ErrorCode.NOT_UNAUTHORIZED_APPLICATION);
+        }
+        application.updateMemo(memoRequestDto);
+        applicationRepository.save(application);
+
     }
 }
